[local_auth] Refactor package to make use of new platform interface and native implementations (#4701)
