+++ packages/local_auth/local_auth/example/lib/main.dart
@@ -79,9 +79,12 @@ class _MyAppState extends State<MyApp> {
         _authorized = 'Authenticating';
       });
       authenticated = await auth.authenticate(
-          localizedReason: 'Let OS determine authentication method',
+        localizedReason: 'Let OS determine authentication method',
+        options: const AuthenticationOptions(
           useErrorDialogs: true,
-          stickyAuth: true);
+          stickyAuth: true,
+        ),
+      );
       setState(() {
         _isAuthenticating = false;
       });
@@ -109,11 +112,14 @@ class _MyAppState extends State<MyApp> {
         _authorized = 'Authenticating';
       });
       authenticated = await auth.authenticate(
-          localizedReason:
-              'Scan your fingerprint (or face or whatever) to authenticate',
+        localizedReason:
+            'Scan your fingerprint (or face or whatever) to authenticate',
+        options: const AuthenticationOptions(
           useErrorDialogs: true,
           stickyAuth: true,
-          biometricOnly: true);
+          biometricOnly: true,
+        ),
+      );
       setState(() {
         _isAuthenticating = false;
         _authorized = 'Authenticating';
