+++ packages/local_auth/local_auth/pubspec.yaml
@@ -3,11 +3,7 @@ description: Flutter plugin for Android and iOS devices to allow local
   authentication via fingerprint, touch ID, face ID, passcode, pin, or pattern.
 repository: https://github.com/flutter/plugins/tree/main/packages/local_auth/local_auth
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+local_auth%22
-version: 1.1.11
-
-# Temporarily disable publishing to allow moving Android and iOS
-# implementations.
-publish_to: none
+version: 2.0.0
 
 environment:
   sdk: ">=2.14.0 <3.0.0"
@@ -24,15 +20,10 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_plugin_android_lifecycle: ^2.0.1
   intl: ^0.17.0
-# Temporary path dependencies to allow moving Android and iOS implementations.
-  local_auth_android:
-    path: ../local_auth_android
-  local_auth_ios:
-    path: ../local_auth_ios
+  local_auth_android: ^1.0.0
+  local_auth_ios: ^1.0.1
   local_auth_platform_interface: ^1.0.1
-  platform: ^3.0.0
 
 dev_dependencies:
   flutter_driver:
@@ -42,3 +33,4 @@ dev_dependencies:
   integration_test:
     sdk: flutter
   mockito: ^5.1.0
+  plugin_platform_interface: ^2.1.2
