+++ packages/local_auth/local_auth/test/local_auth_test.dart
@@ -4,13 +4,14 @@
 
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:local_auth/auth_strings.dart';
 import 'package:local_auth/local_auth.dart';
+import 'package:local_auth/src/local_auth.dart';
+import 'package:local_auth_android/local_auth_android.dart';
+import 'package:local_auth_ios/local_auth_ios.dart';
 import 'package:local_auth_platform_interface/local_auth_platform_interface.dart';
 import 'package:local_auth_platform_interface/types/auth_messages.dart';
 import 'package:local_auth_platform_interface/types/auth_options.dart';
 import 'package:mockito/mockito.dart';
-import 'package:platform/platform.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 void main() {
@@ -22,24 +23,6 @@ void main() {
     localAuthentication = LocalAuthentication();
     mockLocalAuthPlatform = MockLocalAuthPlatform();
     LocalAuthPlatform.instance = mockLocalAuthPlatform;
-  });
-
-  test('authenticateWithBiometrics calls platform implementation', () {
-    when(mockLocalAuthPlatform.authenticate(
-      localizedReason: anyNamed('localizedReason'),
-      authMessages: anyNamed('authMessages'),
-      options: anyNamed('options'),
-    )).thenAnswer((_) async => true);
-    localAuthentication.authenticateWithBiometrics(
-        localizedReason: 'Test Reason');
-    verify(mockLocalAuthPlatform.authenticate(
-      localizedReason: 'Test Reason',
-      authMessages: <AuthMessages>[
-        const IOSAuthMessages(),
-        const AndroidAuthMessages(),
-      ],
-      options: const AuthenticationOptions(biometricOnly: true),
-    )).called(1);
   });
 
   test('authenticate calls platform implementation', () {
@@ -73,18 +56,11 @@ void main() {
     verify(mockLocalAuthPlatform.getEnrolledBiometrics()).called(1);
   });
 
-  test('stopAuthentication calls platform implementation on Android', () {
+  test('stopAuthentication calls platform implementation', () {
     when(mockLocalAuthPlatform.stopAuthentication())
         .thenAnswer((_) async => true);
-    setMockPathProviderPlatform(FakePlatform(operatingSystem: 'android'));
     localAuthentication.stopAuthentication();
     verify(mockLocalAuthPlatform.stopAuthentication()).called(1);
-  });
-
-  test('stopAuthentication does not call platform implementation on iOS', () {
-    setMockPathProviderPlatform(FakePlatform(operatingSystem: 'ios'));
-    localAuthentication.stopAuthentication();
-    verifyNever(mockLocalAuthPlatform.stopAuthentication());
   });
 
   test('canCheckBiometrics returns correct result', () async {
@@ -110,11 +86,8 @@ class MockLocalAuthPlatform extends Mock
 
   @override
   Future<bool> authenticate({
-    String? localizedReason,
-    Iterable<AuthMessages>? authMessages = const <AuthMessages>[
-      IOSAuthMessages(),
-      AndroidAuthMessages()
-    ],
+    required String? localizedReason,
+    required Iterable<AuthMessages>? authMessages,
     AuthenticationOptions? options = const AuthenticationOptions(),
   }) =>
       super.noSuchMethod(
